End of Section 17

+++ client/src/app/components/messages/messages.component.ts
@@ -1,8 +1,10 @@
 import { Component, OnInit } from '@angular/core';
+import { ToastrService } from 'ngx-toastr';
 import { switchAll } from 'rxjs/operators';
 import { Message } from 'src/app/_models/message';
 import { Pagination } from 'src/app/_models/pagination';
 import { MessageService } from 'src/app/_services/message.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-messages',
@@ -17,7 +19,7 @@ export class MessagesComponent implements OnInit {
   pageSize = 5;
   loading = false;
 
-  constructor(private messageService: MessageService) { }
+  constructor(private messageService: MessageService, private toastr: ToastrService) { }
 
   ngOnInit(): void {
     this.loadMessages();
@@ -33,16 +35,57 @@ export class MessagesComponent implements OnInit {
   }
 
   deleteMessage(id: number) {
-    this.messageService.deleteMessage(id).subscribe(() => {
-      this.messages.splice(this.messages.findIndex(m => m.id === id), 1);
+    Swal.fire({
+      title: '<strong><h4>Confirmation.</h4></strong>',
+      icon: 'question',
+      html: 'Please confirm to delete message',
+      showCloseButton: true,
+      showCancelButton: true,
+      confirmButtonText: '<i class="fa fa-thumbs-up"></i> Sure!',
+      confirmButtonColor: 'green',
+      confirmButtonAriaLabel: 'Thumbs up Sure!',
+      cancelButtonText: '<i class="fa fa-thumbs-down"></i> Not Sure!',
+      cancelButtonColor: 'red',
+      allowOutsideClick: false
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.messageService.deleteMessage(id).subscribe(() => {
+          this.messages.splice(this.messages.findIndex(m => m.id === id), 1);
+          Swal.fire({
+            icon: 'success',
+            title: 'Message deleted successfully',
+            showConfirmButton: false,
+            timer: 1500
+          })
+          //   this.swalWithBootstrapButtons.fire(
+          //     'Deleted!',
+          //     'Your message deleted.',
+          //     'success'
+          //   )
+        })
+      } else {
+        this.toastr.error('You have denied to delete message.')
+      }
     })
+    // if(confirm('Confirm to delete message...')){
+    // this.messageService.deleteMessage(id).subscribe(() => {
+    //   this.messages.splice(this.messages.findIndex(m => m.id === id), 1);
+    // })
+    // }
   }
 
+  swalWithBootstrapButtons = Swal.mixin({
+    customClass: {
+      confirmButton: 'btn btn-success',
+      cancelButton: 'btn btn-danger'
+    },
+    buttonsStyling: false
+  })
+
   pageChanged(event: any) {
-    if(this.pageNumber !== event.page) {
+    if (this.pageNumber !== event.page) {
       this.pageNumber = event.page;
     }
     this.loadMessages();
   }
-
 }
